Log the time it takes for a query in es and total
